Reduce pylint false positives on C extensions
Reduce false positive rate of pylint reports due to use of members from
C extension. Whitelist the vulkan module.

Global evaluation
-----------------
Your code has been rated at 8.07/10 (previous run: -14.99/10, +23.07)

Pylint imports modules to effectively identify valid methods and attributes.
It was decided that importing c extensions that are not part of the python
stdlib is a security risk and could introduce malicious code.

This was done in the release of Astroid 1.3.1 [0]

[0] https://mail.python.org/pipermail/code-quality/2014-November/000394.html